chore: remove wcv1 to dApps
