+++ .env.base
@@ -14,7 +14,6 @@ REACT_APP_FEATURE_MIXPANEL=true
 REACT_APP_FEATURE_OPTIMISM=true
 REACT_APP_FEATURE_SAVERS_VAULTS=true
 REACT_APP_FEATURE_TRADE_RATES=true
-REACT_APP_FEATURE_WALLET_CONNECT_TO_DAPPS=true
 REACT_APP_FEATURE_WALLET_CONNECT_TO_DAPPS_V2=true
 REACT_APP_FEATURE_WHEREVER=true
 REACT_APP_FEATURE_YAT=true
