+++ src/AppProviders.tsx
@@ -5,7 +5,6 @@ import {
   createStandaloneToast,
 } from '@chakra-ui/react'
 import { DefiManagerProvider } from 'features/defi/contexts/DefiManagerProvider/DefiManagerProvider'
-import { WalletConnectBridgeProvider } from 'plugins/walletConnectToDapps/v1/WalletConnectBridgeProvider'
 import { WalletConnectV2Provider } from 'plugins/walletConnectToDapps/v2/WalletConnectV2Provider'
 import React, { useCallback } from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
@@ -56,23 +55,21 @@ export function AppProviders({ children }: ProvidersProps) {
                 <BrowserRouterProvider>
                   <I18nProvider>
                     <WalletProvider>
-                      <WalletConnectBridgeProvider>
-                        <WalletConnectV2Provider>
-                          <KeepKeyProvider>
-                            <ErrorBoundary FallbackComponent={ErrorPage} onError={handleError}>
-                              <ModalProvider>
-                                <TransactionsProvider>
-                                  <AppProvider>
-                                    <FoxEthProvider>
-                                      <DefiManagerProvider>{children}</DefiManagerProvider>
-                                    </FoxEthProvider>
-                                  </AppProvider>
-                                </TransactionsProvider>
-                              </ModalProvider>
-                            </ErrorBoundary>
-                          </KeepKeyProvider>
-                        </WalletConnectV2Provider>
-                      </WalletConnectBridgeProvider>
+                      <WalletConnectV2Provider>
+                        <KeepKeyProvider>
+                          <ErrorBoundary FallbackComponent={ErrorPage} onError={handleError}>
+                            <ModalProvider>
+                              <TransactionsProvider>
+                                <AppProvider>
+                                  <FoxEthProvider>
+                                    <DefiManagerProvider>{children}</DefiManagerProvider>
+                                  </FoxEthProvider>
+                                </AppProvider>
+                              </TransactionsProvider>
+                            </ModalProvider>
+                          </ErrorBoundary>
+                        </KeepKeyProvider>
+                      </WalletConnectV2Provider>
                     </WalletProvider>
                   </I18nProvider>
                 </BrowserRouterProvider>
