+++ src/plugins/walletConnectToDapps/components/modals/connect/Connect.tsx
@@ -1,8 +1,5 @@
 import { Modal, ModalCloseButton, ModalContent, ModalOverlay } from '@chakra-ui/react'
-import { fromAccountId } from '@shapeshiftoss/caip'
 import { ConnectContent } from 'plugins/walletConnectToDapps/components/modals/connect/ConnectContent'
-import { isWalletConnectV2Uri } from 'plugins/walletConnectToDapps/components/modals/connect/utils'
-import { useWalletConnect } from 'plugins/walletConnectToDapps/v1/WalletConnectBridgeContext'
 import { useWalletConnectV2 } from 'plugins/walletConnectToDapps/v2/WalletConnectV2Provider'
 import { useCallback } from 'react'
 
@@ -17,17 +14,7 @@ const minWidthProp = { base: '100%', md: '500px' }
 const maxWidthProp = { base: 'full', md: '500px' }
 
 const Connect = ({ initialUri, isOpen, onClose }: Props) => {
-  const { connect, wcAccountId } = useWalletConnect()
   const { pair } = useWalletConnectV2()
-
-  const handleConnectV1 = useCallback(
-    (uri: string) => {
-      if (!wcAccountId) return
-      const connectionResult = connect(uri, fromAccountId(wcAccountId).account)
-      if (connectionResult?.successful) onClose()
-    },
-    [connect, onClose, wcAccountId],
-  )
 
   const handleConnectV2 = useCallback(
     async (uri: string) => {
@@ -35,15 +22,6 @@ const Connect = ({ initialUri, isOpen, onClose }: Props) => {
       if (connectionResult) onClose()
     },
     [onClose, pair],
-  )
-
-  const handleConnect = useCallback(
-    (uri: string) => {
-      // https://eips.ethereum.org/EIPS/eip-1328
-      const isWalletConnectV2 = isWalletConnectV2Uri(uri)
-      return isWalletConnectV2 ? handleConnectV2(uri) : handleConnectV1(uri)
-    },
-    [handleConnectV1, handleConnectV2],
   )
 
   return (
@@ -58,7 +36,7 @@ const Connect = ({ initialUri, isOpen, onClose }: Props) => {
         maxWidth={maxWidthProp}
       >
         <ModalCloseButton position='absolute' color='text.subtle' />
-        <ConnectContent initialUri={initialUri} handleConnect={handleConnect} />
+        <ConnectContent initialUri={initialUri} handleConnect={handleConnectV2} />
       </ModalContent>
     </Modal>
   )
