+++ src/plugins/walletConnectToDapps/types.ts
@@ -1,3 +1,5 @@
+import type { FeeDataKey } from '@shapeshiftoss/chain-adapters'
+
 export type RegistryItem = {
   category: string
   id: string
@@ -6,12 +8,24 @@ export type RegistryItem = {
   image: string
 }
 
-export type APIRegistryItem = {
-  app_type: string
-  id: string
-  homepage: string
-  name: string
-  image_url: {
-    md: string
+export type ConfirmData = {
+  nonce?: string
+  gasLimit?: string
+  speed: WalletConnectFeeDataKey
+  customFee?: {
+    baseFee: string
+    priorityFee: string
   }
 }
+
+export type TransactionParams = {
+  from: string
+  to: string
+  data: string
+  gas?: string
+  gasPrice?: string
+  value?: string
+  nonce?: string
+}
+
+export type WalletConnectFeeDataKey = FeeDataKey | 'custom'
