--- src/plugins/walletConnectToDapps/v1/components/modals/callRequest/CallRequestCommon.ts
@@ -1,12 +1,0 @@
-import type { FeeDataKey } from '@shapeshiftoss/chain-adapters'
-
-export type WalletConnectFeeDataKey = FeeDataKey | 'custom'
-export type ConfirmData = {
-  nonce?: string
-  gasLimit?: string
-  speed: WalletConnectFeeDataKey
-  customFee?: {
-    baseFee: string
-    priorityFee: string
-  }
-}
