+++ src/plugins/walletConnectToDapps/v2/types.ts
@@ -2,7 +2,7 @@ import type { ChainId } from '@shapeshiftoss/caip'
 import type { ICore, SessionTypes, SignClientTypes } from '@walletconnect/types'
 import type { PairingTypes } from '@walletconnect/types/dist/types/core/pairing'
 import type { IWeb3Wallet, Web3WalletTypes } from '@walletconnect/web3wallet'
-import type { WalletConnectFeeDataKey } from 'plugins/walletConnectToDapps/v1/components/modals/callRequest/CallRequestCommon'
+import type { WalletConnectFeeDataKey } from 'plugins/walletConnectToDapps/types'
 import type { Dispatch } from 'react'
 import type { PartialRecord } from 'lib/utils'
